move doremi to /examples
