+++ examples/doremi/train_doremi.py
@@ -9,11 +9,12 @@ torchrun --nproc_per_node=4 examples/doremi/train_doremi.py --config-file exampl
 import argparse
 
 import torch
+from doremi.config import DoReMiConfig
+from doremi.dataloader import get_dataloader, get_datasets
+from doremi.trainer import DoReMiTrainer
+from doremi.utils import compute_domain_weights_based_on_token_count
+
 from nanotron.config import get_config_from_file
-from nanotron.doremi.config import DoReMiConfig
-from nanotron.doremi.dataloader import get_dataloader, get_datasets
-from nanotron.doremi.trainer import DoReMiTrainer
-from nanotron.doremi.utils import compute_domain_weights_based_on_token_count
 
 
 def get_args():
