+++ src/nanotron/trainer.py
@@ -6,9 +6,20 @@ import time
 from dataclasses import asdict
 from pathlib import Path
 from pprint import pformat
-from typing import Callable, Dict, Iterable, Iterator, List, Optional, Tuple, Type, Union
+from typing import (
+    Callable,
+    Dict,
+    Iterable,
+    Iterator,
+    List,
+    Optional,
+    Tuple,
+    Type,
+    Union,
+)
 
 import torch
+import wandb
 from torch.nn.parallel import DistributedDataParallel
 
 from nanotron import distributed as dist
@@ -29,7 +40,13 @@ from nanotron.helpers import (
     log_throughput,
     lr_scheduler_builder,
 )
-from nanotron.logging import LoggerWriter, LogItem, human_format, log_rank, set_logger_verbosity_format
+from nanotron.logging import (
+    LoggerWriter,
+    LogItem,
+    human_format,
+    log_rank,
+    set_logger_verbosity_format,
+)
 from nanotron.models import NanotronModel, build_model
 from nanotron.models.base import check_model_has_grad
 from nanotron.models.llama import LlamaForTraining, RotaryEmbedding
@@ -38,9 +55,7 @@ from nanotron.optim.clip_grads import clip_grad_norm
 from nanotron.parallel import ParallelContext
 from nanotron.parallel.data_parallel.utils import sync_gradients_across_dp
 from nanotron.parallel.parameters import NanotronParameter, sanity_check
-from nanotron.parallel.pipeline_parallel.engine import (
-    PipelineEngine,
-)
+from nanotron.parallel.pipeline_parallel.engine import PipelineEngine
 from nanotron.parallel.pipeline_parallel.tensor_pointer import TensorPointer
 from nanotron.parallel.pipeline_parallel.utils import get_pp_rank_of
 from nanotron.parallel.tensor_parallel.nn import (
@@ -53,9 +68,7 @@ from nanotron.parallel.tied_parameters import (
     sync_tied_weights_gradients,
     tie_parameters,
 )
-from nanotron.random import (
-    set_random_seed,
-)
+from nanotron.random import set_random_seed
 from nanotron.sanity_checks import (
     after_optim_step_sanity_checks,
     after_tbi_sanity_checks,
@@ -223,7 +236,13 @@ class DistributedTrainer:
         pass
 
     def pre_training(self, *args, **kwargs):
-        pass
+        current_time = datetime.datetime.now().strftime("%d/%m/%Y_%H:%M:%S")
+        if dist.get_rank(self.parallel_context.world_pg) == 0:
+            wandb.init(
+                project=self.config.general.project,
+                name=f"{current_time}_{self.config.general.project}_{self.config.general.run}",
+                config={"version": 1, "nanotron_config": self.config.as_dict()},
+            )
 
     def post_train_step(self):
         pass
@@ -483,6 +502,10 @@ class DistributedTrainer:
                         LogItem("hd_free_memory_tb", free, "human_format"),  #  / (2**40), ".2f"),
                     ]
                 )
+
+            wandb.log(
+                {**{log_item.tag: log_item.scalar_value for log_item in log_entries}, "step": self.iteration_step}
+            )
 
             self.loggerwriter.add_scalars_from_list(log_entries, self.iteration_step)
